Phase 4: split CodeMirrorEditor into eight modules
CodeMirrorEditor.tsx was 993 lines, and only the last 140 of them were the
React component. The rest was CodeMirror wiring — widgets, state fields,
view plugins, extension factories — that has nothing to do with React and
was unreachable from a test.

Split along what each piece configures:

- codeMirrorTypes.ts       the four shared types
- codeMirrorLanguage.ts    language -> Extension
- codeMirrorColorPicker.ts ColorSwatchWidget + its decoration plugin
- codeMirrorTheme.ts       darkHighlight + darkChrome
- codeMirrorDiagnostics.ts server diagnostics + the advisory ts lint source
- codeMirrorHover.ts       modifier-aware hover, goto handler
- codeMirrorGhostText.ts   TA-02 ghost text end to end
- codeMirrorCompletion.ts  measured ts autocomplete + visibility tracking

CodeMirrorEditor.tsx keeps its props, languageServiceExtensions (the one
place those pieces are composed), and the component: 221 lines.

CodeMirrorDiagnostic and CodeMirrorLanguageService now live in
codeMirrorTypes.ts and are re-exported from CodeMirrorEditor.tsx, so
CodeSurface.tsx's type import is unchanged. FetchGhostText and
GotoDefinitionHandler were duplicated inline in the props type and in the
ghost-text block; both now name the shared type.

No behaviour change: same extensions, composed in the same order.
