+++ apps/web/src/surfaces/studio/codeMirrorTheme.ts
@@ -1,0 +1,166 @@
+import { HighlightStyle } from '@codemirror/language';
+import { EditorView } from '@codemirror/view';
+import { tags } from '@lezer/highlight';
+
+// Palette matches the read-only viewer's .code-tok-* colors.
+export const darkHighlight = HighlightStyle.define([
+  { tag: [tags.comment, tags.lineComment, tags.blockComment], color: 'var(--muted, #8b949e)' },
+  { tag: [tags.string, tags.special(tags.string), tags.regexp], color: '#7ee787' },
+  { tag: [tags.number, tags.bool, tags.null, tags.atom], color: '#79c0ff' },
+  { tag: [tags.keyword, tags.operatorKeyword, tags.modifier, tags.self], color: '#ff7b72' },
+  { tag: [tags.propertyName, tags.attributeName, tags.labelName], color: '#d2a8ff' },
+  { tag: [tags.function(tags.variableName), tags.function(tags.propertyName)], color: '#d2a8ff' },
+  { tag: [tags.typeName, tags.className, tags.namespace], color: '#ffa657' },
+  { tag: [tags.definition(tags.variableName), tags.variableName], color: 'var(--text, #e6edf3)' },
+  { tag: [tags.operator, tags.punctuation, tags.separator, tags.bracket], color: 'var(--text, #e6edf3)' },
+  { tag: tags.heading, color: '#79c0ff', fontWeight: 'bold' },
+  { tag: tags.link, color: '#7ee787', textDecoration: 'underline' },
+  { tag: tags.invalid, color: '#ff7b72', textDecoration: 'underline wavy' },
+]);
+
+export const darkChrome = EditorView.theme(
+  {
+    '&': { height: '100%', fontSize: '0.84rem', backgroundColor: 'transparent', color: 'var(--text, #e6edf3)' },
+    '.cm-scroller': { fontFamily: 'var(--mono-font, monospace)', overflow: 'auto' },
+    '.cm-content': { caretColor: '#00e4ac' },
+    '.cm-cursor, .cm-dropCursor': { borderLeftColor: '#00e4ac' },
+    '.cm-activeLine': { backgroundColor: 'rgba(148, 163, 184, 0.07)' },
+    '.cm-activeLineGutter': { backgroundColor: 'rgba(148, 163, 184, 0.07)' },
+    '&.cm-focused > .cm-scroller > .cm-selectionLayer .cm-selectionBackground, .cm-selectionBackground': {
+      backgroundColor: 'rgba(0, 228, 172, 0.16)',
+    },
+    '.cm-selectionMatch': { backgroundColor: 'rgba(0, 228, 172, 0.12)' },
+    '.cm-searchMatch': { backgroundColor: 'rgba(56, 189, 248, 0.22)', borderRadius: '2px' },
+    '.cm-searchMatch.cm-searchMatch-selected': { backgroundColor: 'rgba(0, 228, 172, 0.38)' },
+    '&.cm-focused .cm-matchingBracket': { backgroundColor: 'rgba(0, 228, 172, 0.2)', outline: 'none' },
+    '&.cm-focused .cm-nonmatchingBracket': { backgroundColor: 'rgba(255, 123, 114, 0.2)' },
+    '.cm-gutters': { backgroundColor: 'transparent', color: 'var(--muted, #8b949e)', border: 'none' },
+    '.cm-foldGutter, .cm-lineNumbers': { color: 'var(--muted, #8b949e)' },
+    '.cm-tooltip': {
+      backgroundColor: 'var(--panel, #161c22)',
+      color: 'var(--text, #e6edf3)',
+      border: '1px solid rgba(148, 163, 184, 0.25)',
+      borderRadius: '8px',
+    },
+    '.cm-tooltip.cm-tooltip-autocomplete > ul > li[aria-selected]': {
+      backgroundColor: 'rgba(0, 228, 172, 0.16)',
+      color: 'var(--text, #e6edf3)',
+    },
+    // The search widget floats over the editor, so panels reserve no space.
+    '.cm-panels': { backgroundColor: 'transparent', color: 'var(--text, #e6edf3)' },
+    '.cm-panels.cm-panels-top': {
+      position: 'absolute',
+      top: 0,
+      right: 0,
+      left: 'auto',
+      zIndex: 12,
+      borderBottom: 'none',
+    },
+    '.cm-vs-search': {
+      display: 'flex',
+      alignItems: 'flex-start',
+      gap: '2px',
+      margin: '6px 16px 0 6px',
+      padding: '6px 6px 6px 2px',
+      borderRadius: '10px',
+      border: '1px solid var(--panel-border, rgba(148, 163, 184, 0.25))',
+      backgroundColor: 'var(--panel, #161c22)',
+      boxShadow: '0 10px 30px rgba(0, 0, 0, 0.45)',
+      fontSize: '0.76rem',
+      maxWidth: 'calc(100% - 22px)',
+    },
+    '.cm-vs-body': { display: 'flex', flexDirection: 'column', gap: '4px', minWidth: 0 },
+    // Wraps rather than overflowing the editor on a narrow viewport.
+    '.cm-vs-row': { display: 'flex', flexWrap: 'wrap', alignItems: 'center', gap: '2px', minWidth: 0 },
+    // Without this, display:flex beats the hidden attribute and replace never collapses.
+    '.cm-vs-row[hidden]': { display: 'none' },
+    // The chevron spans both rows, as in VS Code.
+    '.cm-vs-expand': { alignSelf: 'stretch', height: 'auto', flex: 'none' },
+    '.cm-vs-expanded .cm-vs-expand svg': { transform: 'rotate(90deg)' },
+    '.cm-vs-expand svg': { transition: 'transform 120ms ease' },
+    '.cm-vs-field-wrap': {
+      display: 'flex',
+      alignItems: 'center',
+      gap: '2px',
+      // Fixed, not growing, so both fields match width.
+      flex: '0 1 15rem',
+      width: '15rem',
+      minWidth: '6rem',
+      padding: '0 3px 0 7px',
+      borderRadius: '6px',
+      border: '1px solid var(--panel-border, rgba(148, 163, 184, 0.25))',
+      backgroundColor: 'rgba(9, 13, 17, 0.7)',
+    },
+    '.cm-vs-field-wrap:focus-within': {
+      borderColor: 'var(--turquoise, #00e4ac)',
+      boxShadow: '0 0 0 1px rgba(0, 228, 172, 0.35)',
+    },
+    '.cm-vs-invalid .cm-vs-field-wrap:first-of-type': { borderColor: 'var(--error, #ff6b6b)' },
+    '.cm-vs-field': {
+      flex: 1,
+      minWidth: 0,
+      border: 'none',
+      outline: 'none',
+      background: 'transparent',
+      color: 'var(--text, #e6edf3)',
+      fontFamily: 'var(--mono-font, monospace)',
+      fontSize: '0.76rem',
+      padding: '5px 0',
+    },
+    '.cm-vs-field::placeholder': { color: 'var(--muted, #8b949e)' },
+    '.cm-vs-toggles': { display: 'flex', alignItems: 'center', gap: '1px', flex: 'none' },
+    '.cm-vs-toggle': {
+      display: 'flex',
+      alignItems: 'center',
+      justifyContent: 'center',
+      width: '20px',
+      height: '20px',
+      padding: 0,
+      border: 'none',
+      borderRadius: '4px',
+      background: 'transparent',
+      color: 'var(--muted, #8b949e)',
+      font: 'inherit',
+      fontSize: '0.68rem',
+      lineHeight: 1,
+      cursor: 'pointer',
+    },
+    '.cm-vs-toggle:hover': { backgroundColor: 'rgba(148, 163, 184, 0.16)', color: 'var(--text, #e6edf3)' },
+    '.cm-vs-toggle[aria-pressed=true]': {
+      backgroundColor: 'rgba(0, 228, 172, 0.18)',
+      color: 'var(--turquoise, #00e4ac)',
+    },
+    '.cm-vs-count': {
+      flex: 'none',
+      minWidth: '4.6rem',
+      padding: '0 4px',
+      color: 'var(--muted, #8b949e)',
+      fontSize: '0.7rem',
+      whiteSpace: 'nowrap',
+      textAlign: 'right',
+    },
+    '.cm-vs-icon-button': {
+      display: 'flex',
+      alignItems: 'center',
+      justifyContent: 'center',
+      flex: 'none',
+      width: '24px',
+      height: '24px',
+      padding: 0,
+      border: 'none',
+      borderRadius: '5px',
+      background: 'transparent',
+      color: 'var(--muted, #8b949e)',
+      font: 'inherit',
+      cursor: 'pointer',
+    },
+    '.cm-vs-text-button': { width: 'auto', padding: '0 7px', fontSize: '0.72rem', fontWeight: 500 },
+    '.cm-vs-icon-button:hover': { backgroundColor: 'rgba(148, 163, 184, 0.16)', color: 'var(--text, #e6edf3)' },
+    '.cm-vs-icon-button:active': { backgroundColor: 'rgba(0, 228, 172, 0.2)' },
+    '.cm-vs-search button:focus-visible': {
+      outline: '1px solid var(--turquoise, #00e4ac)',
+      outlineOffset: '-1px',
+    },
+  },
+  { dark: true },
+);
